Working on automatically building the OpenCR firmware
A rosrun command will let the firmware be rebuilt.

The firmware itself is coupled to this package, because it uses
rosserial, and we must generater the arduino-style messages for
the custom raw messages.

Therefore, the ultimate workflow will be to do this all at once
1. Install all ROS dependencies with rosdep
2. catkin_build the nuturtlebot package
3. rosrun a build script that will
    - fetch the arduino environment
    - update the rosserial library with what our current system has
      (and the current system should be at the same version as the
turtlebot image
    - build the custom code with up-to-date messages.

Building the firmware is a separate afirmative step because
it downloads dependencies and is not needed for the usual usage of this
package.  It just makes sense to couple it with nuturtlebot since
the custom firmware needs to be updated.

A less-hacky solution would be to convert the opencr package into
a ROS package and build everything in one workspace, but
there is not enough time right now and this should be an improvement