Properly rebuilt the rosserial files.
The build script is now very hacky:
it checks out two branches, deletes the rosserial files
and regenerates them.

Really, the rosserial files should not be part of the repository in the
first place, since they are a dependency.  But I keep them in the repo
to make updating it easier.

The one downside is this package has some dependencies that need
to be sourced for this to work